+++ js/data.js
@@ -1,0 +1,50 @@
+import {getRandomElement} from './util.js';
+
+const POSTS = 25;
+const MIN_LIKES_COUNT = 15;
+const MAX_LIKES_COUNT = 200;
+const COMMENTS_COUNT = 30;
+const AVATARS_COUNT = 6;
+
+const DESCRIPTIONS = [
+  'Красная роза в утренних каплях росы',
+  'Нежный белый тюльпан на фоне зелёной лужайки',
+  'Фиолетовый ирис в солнечном свете полудня',
+  'Жёлтая хризантема среди осенних листьев',
+  'Пышные розовые пионы в саду',
+  'Оранжевая лилия на фоне заката',
+  'Голубой василёк посреди поля',
+  'Белая камелия на ветке',
+  'Романтичный букет сирени',
+  'Золотистые нарциссы под ясным небом',
+  'Сиреневые крокусы в весеннем парке',
+  'Цветущие магнолии в лучах рассвета',
+  'Малиновый гвоздик в зелёной траве',
+  'Кувшинка на тихой воде',
+  'Фуксийнные азалии в красивом саду',
+  'Лавандовое поле под голубым небом',
+  'Полевой мак на ветреном поле',
+  'Розовые орхидеи в прозрачной вазе',
+  'Белоснежный жасмин на дереве',
+  'Фрезии разных оттенков в саду',
+  'Фиолетовый лотос на пруду',
+  'Живописные георгины на клумбе',
+  'Бордовые астры на фоне забора',
+  'Солнечные подсолнухи в августе',
+  'Нежные анемоны в букете',
+];
+
+const COMMENTATORS_NAMES = [
+  'Алексей', 'Мария', 'Игорь', 'Екатерина', 'Дмитрий', 'Анна', 'Владимир', 'Ольга', 'Иван', 'Юлия', 'Николай', 'Светлана', 'Сергей', 'Наталья', 'Андрей', 'Елена', 'Александр', 'Татьяна', 'Михаил', 'Алёна', 'Виктор', 'Ирина', 'Павел', 'Евгения', 'Роман', 'Людмила', 'Константин', 'Вероника', 'Максим', 'Алиса'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+
